requirements.txt
aiogram==2.8
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
emoji==0.5.4
idna==2.9
multidict==4.7.6
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
sentry-sdk==0.14.4
six==1.15.0
urllib3==1.25.9
yarl==1.4.2
